+++ app/src/main/java/eu/ase/proiect/util/User.java
@@ -1,16 +1,15 @@
 package eu.ase.proiect.util;
 
-import android.content.Intent;
-
 import java.util.HashMap;
-import java.util.List;
 import java.util.Map;
+
+import eu.ase.proiect.database.model.Book;
 
 public class User {
 
     private int idUser;
     public static String name;
-    public static Map<Integer,Book> mapFavoriteBook= new HashMap<Integer, Book>();
+    public static Map<Long, Book> mapFavoriteBook= new HashMap<Long, Book>();
     public static Map<Integer,Book> mapBookRead = new HashMap<Integer, Book>();;
 
 //cancer
